Update memory.py
I modified here @muellerzr suggests passing `accelerator._schedulers` as a bool value to indicate whether the learning rate schedulers need to be updated based on the current batch size.
The `accelerator._schedulers` attribute in Accelerate refers to a flag that controls whether the learning rate schedulers should be updated or not. By passing it as a bool value, you can make the modification more flexible and configurable.