quant_state of Params4bit became None after "to" method
The **kwargs was missing in line 401; and previously the new class of bitsandbytes Params4bits was not included in line 389, which leads to the missing of quant_state after moving to another device because the re-initialization in line 401 will lead to a default quant_state=None due to lack of **kwargs. The processing of Params4bit is now moved to line 399, but I also added **kwargs to line 401 in case.